Bug 24812: (bug 24081 follow-up) Add permission description
Bug 24081 added a discount permission to the accounts system but during
subsequent rebases, the description of that new permission got lot.

This patch restores the description.

Test plan
0/ Check that a checkbox without a description appears in the 'Manage
patrons patrons fines and fees' section of the 'Set permissions' page.
1/ Apply the patch
2/ Reload the page and note that the checkbox now preceeds a suitable
description.
3/ Signoff.

Bonus points: Check the permission works as described and hides the
'Discount' action button for staff who do not have the permission
enabled.

Signed-off-by: Michal Denar <black23@gmail.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>