Bug 24378: Change wording of auto_renewals notice in updatedatabase
On the template of AUTO-RENEWALS the wording is grammatically incorrect:

[% IF checkout.auto_renew_error == 'too_many' %]
You have reach the maximum of checkouts possible.

This should say
[% IF checkout.auto_renew_error == 'too_many' %]
You have reached the maximum of checkouts possible.

Test plan:
- Make sure you have the AUTO_RENEWALS notice with the problematic
  spelling
- install patch
- update database
- confirm changes to notice name and wording

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Kept Andrew's change on the original updatedatabase entry,
but removed the database update to be done a little differently
in a follow-up patch.

Amended test plan.

Signed-off-by: Andrew Fuerste-Henry <andrew@bywatersolutions.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>